refactored: restore_mnemonic_backup
