+++ lib/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/export.dart
@@ -1,1 +1,3 @@
 export 'post_details_reactions_list/export.dart';
+export 'post_comments_list/export.dart';
+export 'post_details_bottom_bar/export.dart';
